feat: implement dynamic metadata generation for blog posts, enhancing SEO and social sharing capabilities with Open Graph and Twitter card support
